Update content to new overlay format
