+++ koerier/src/index.ts
@@ -1,4 +1,5 @@
 import { default as world } from "./game";
 import "./locations/index.ts";
+import "./conversations/dwarf.ts";
 
 export default world;
